refactor: move navs as internal components or partials for AppLayout
