--- src/components/nav/Footer.tsx
@@ -1,3 +1,0 @@
-export function Footer() {
-  return <footer className="w-screen py-4 px-7">Empty footer</footer>;
-}
