--- src/components/nav/navLinks.ts
@@ -1,4 +1,0 @@
-export const navLinks = [
-  { label: 'Stake', to: '/' },
-  { label: 'Unstake', to: '/unstake' },
-];
