--- src/layout/HeadMeta.tsx
@@ -1,22 +1,0 @@
-import Head from 'next/head';
-
-function toTitleCase(str: string) {
-  return str.replace(/\w\S*/g, (txt) => {
-    return txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase();
-  });
-}
-
-export function HeadMeta({ pathName }: { pathName: string }) {
-  return (
-    <Head>
-      <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-      <title>{`Liquid Staking | ${getHeadTitle(pathName)}`}</title>
-    </Head>
-  );
-}
-
-function getHeadTitle(pathName: string) {
-  const segments = pathName.split('/');
-  if (segments.length <= 1 || !segments[1]) return 'Home';
-  else return toTitleCase(segments[1]);
-}
